feat: add useTrackingDetails as sharable component and update it to use Modal instead of Overlay
